fix: other pages with percentages
